requirements.txt
# Core data science
pandas>=2.2.3
numpy>=2.1.0
scikit-learn>=1.5.0
imbalanced-learn>=0.12.0   # For SMOTE

# Model training
xgboost>=2.0.3
lightgbm>=4.3.0

# Experiment tracking
mlflow>=2.11.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.28.0
httpx>=0.27.0              # Required by FastAPI TestClient

# Data versioning
dvc>=3.48.4
dvc-s3>=3.2.0              # DVC plugin for AWS S3

# Visualisation
matplotlib>=3.8.2
seaborn>=0.13.2

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0          # Test coverage reports

# Utilities
joblib>=1.3.2               # Model serialisation
pyyaml>=6.0.1               # For reading params.yaml
shap>=0.45.0                # SHAP explainability

# Monitoring (Week 5)
prometheus-fastapi-instrumentator>=0.9.0   # Auto RED metrics for FastAPI at /prometheus
evidently>=0.4.33                          # ML observability — data-drift reports

# Orchestration (Week 6)
prefect>=2.19.0              # Python-native workflow orchestrator

# Load testing (Week 8)
locust>=2.20.0               # Python-native load testing framework
